# openapi.yml
openapi: 3.0.0
info:
  title: School Management System API
  version: 1.0.0
  description: API documentation for School Management System
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.production.com
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          default: false
          example: false
        message:
          type: string
          example: An error occurred
      required:
        - success
        - message
...
    Exam:
      type: object
      properties:
        examId:
          type: number
        examName:
          type: string
        academicYear:
          type: string
        classId:
          type: number
        sectionId:
          type: number
        examDate:
          type: string
          format: date
        totalMarks:
          type: number
        passingMarks:
          type: number
        examType:
          type: string
        description:
          type: string
        isPublished:
          type: boolean
      required:
        - examName
        - classId
        - examDate
        - academicYear
        - totalMarks
    ExamResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Exam'
        message:
          type: string
...
    ResultResponse:
      type: object
      properties:
        resultId:
          type: integer
          description: ID of the result.
          example: 1
        studentId:
          type: integer
          description: ID of the student.
          example: 1
        examId:
          type: integer
          description: ID of the exam.
          example: 10
        subjectId:
          type: integer
          description: ID of the subject.
          example: 5
        marksObtained:
          type: number
          description: Marks obtained by the student.
          example: 85.5
      required:
        - studentId
        - examId
        - subjectId
        - marksObtained
...
security:
  - bearerAuth: []
paths:
...
  /exams:
    post:
      summary: Create a new exam
      tags:
        - Exams
      security:
        - bearerAuth: []
      requestBody:
        description: Exam creation data
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Exam'
            examples:
              exam:
                summary: Exam creation payload
                value:
                  examName: Midterm Exam
                  classId: 1
                  examDate: '2023-07-15'
                  academicYear: 2023-2024
                  totalMarks: 100
                  passingMarks: 40
                  examType: Unit Test
                  description: Midterm examination for 10th grade students
                  isPublished: false
      responses:
        '201':
          description: Exam created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExamResponse'
        '400':
          description: Invalid input data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: Fetch all exams
      tags:
        - Exams
      responses:
        '200':
          description: List of exams retrieved successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Exam'
        '500':
          description: Server error retrieving exams.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /exams/{id}:
    get:
      summary: Fetch an exam by ID
      tags:
        - Exams
      parameters:
        - in: path
          name: id
          description: Numeric ID of the exam to fetch
          required: true
          schema:
            type: number
      responses:
        '200':
          description: Exam retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exam'
        '400':
          description: Invalid exam ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Exam not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Update an exam
      tags:
        - Exams
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          description: Numeric ID of the exam to update
          required: true
          schema:
            type: number
      requestBody:
        description: Updated exam data
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Exam'
            examples:
              exam:
                summary: Updated exam payload
                value:
                  examName: Updated Midterm Exam
                  classId: 1
                  examDate: '2023-07-20'
                  academicYear: 2023-2024
                  totalMarks: 100
                  passingMarks: 40
                  examType: Unit Test
                  description: Updated information for midterm exam
                  isPublished: false
      responses:
        '200':
          description: Exam updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exam'
        '400':
          description: Invalid exam ID or input data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Exam not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete an exam
      tags:
        - Exams
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          description: Numeric ID of the exam to delete
          required: true
          schema:
            type: number
      responses:
        '204':
          description: Exam deleted successfully. No content returned.
        '400':
          description: Invalid exam ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Exam not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /results/exam/{examId}:
    get:
      tags:
        - Results
      summary: Get results by exam ID
      description: Retrieves all exam results for a specific exam.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: examId
          required: true
          description: ID of the exam to retrieve results for
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation - returns list of results for the exam
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResultResponse'
        '400':
          description: Bad request - invalid exam ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...